order-api/requirements.txt
uvicorn==0.13.3
fastapi==0.63.0
elasticsearch==7.10.1
pymemcache==3.4.0
pytest==6.2.2
psycopg2==2.8.6
psycopg2-binary==2.8.6
python-dotenv==0.15.0    
requests==2.25.1